Move eventMapFragment to the .ui.eventShowcase.fragments package
