+++ app/src/main/res/layout/fragment_event_map.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:layout_marginTop="?attr/actionBarSize"
-    tools:context=".ui.eventShowcase.EventMapFragment"
+    tools:context=".ui.eventShowcase.fragments.EventMapFragment"
     android:id="@+id/mapFragment">
 
 </android.support.constraint.ConstraintLayout>
